docker-compose/auth-nfc/requirements.txt
Flask
Flask-CORS
PyJWT
psycopg2-binary